+++ src/Entities/enemies.js
@@ -1,6 +1,6 @@
 import Matter from 'matter-js';
 import Constants, { GetAbsolutHeightPosition, GetAbsolutWidthPosition } from '../Constants';
-import EnemyPlayer from './enemyPlayer';
+import EnemyPlayer from './EnemyPlayer';
 import { skinColor } from '../Constants/skinColor';
 
 export function addEnemies(engine, club) {
